Abe/Heatmap
Abe now processes stuff on the server, so we're not leaking our api key to the client.
Additionally, his family secret is now stored in a .env file so if we wish to stop
broadcasting our api key in our repo, which is acknowledged by wizards as certainly evil,
we can do that someday.

Heatmap now works.
pip install -r tools/requirements.txt
python tools/generate_heatmap.py

Abe takes commands to start the heatmap because I figured it would be impossible to
accidentally start/stop and it wouldn't require a ton of mouse movements. And I'm lazy.

recheatmap and dlheatmap in Abe for recording and downloading the heatmap respectively.
NOTE that dlheatmap does not clear the heatmap buffer or stop recording. Please reload the
page between users/tests.